refactor with last comments and Utils component added
