--- components/exercises/chapter-4-oscar/arraySum.test.tsx
@@ -1,16 +1,0 @@
-import { arraySum } from './arraySum';
-
-describe('arraySum', () => {
-  it('should return 0 for an empty array', () => {
-    expect(arraySum([])).toBe(0);
-  });
-
-  it('should return the sum of a single-element array', () => {
-    expect(arraySum([5])).toBe(5);
-  });
-
-  it('should return the sum of a multi-element array', () => {
-    expect(arraySum([1, 2, 3, 4, 5])).toBe(15);
-    expect(arraySum([-1, -2, -3, -4, -5])).toEqual(-15);
-  });
-});
