--- components/exercises/chapter-4-oscar/arraySum.tsx
@@ -1,10 +1,0 @@
-/**
- * Calculates the sum of all numbers in an array.
- * @param arr - The array of numbers to sum.
- * @returns The sum of all numbers in the array.
- */
-export const arraySum = (arr: number[]): number => {
-  let sum = 0;
-  for (let i = 0; i < arr.length; i++) sum += arr[i];
-  return sum;
-};
